Various fixes, improvements and version bump (#14)
* notification_proxy: Fix np_observe_notifications

Use a correct method for passing strings to C using a vector.
End a vector with a null pointer as the underlying function requires.

* file_relay: Fix a memory leak & null ptr

Use of CString::into_raw() causes a memory leak if a pointer is not
converted back to a CString. We can avoid all of that using two vectors that
get dropped automatically. A vector should also end with a null as the underlying
function requires.

* service: Fix an invalid vector pointer

A vector should be allocated before transferring it to the C side.
Truncate a vector in the end using `received` as a length.

* debug_server: Fix a memory leak

Use a two vector method instead of CString::into_raw()

* instproxy: Use into_iter instead of iter

to avoid cloning objects strings, small code refactoring

* instproxy: Refactor some code

Combine two interators into one, use map_or

* mobile_image_mounter: Refactor some code

Remove unnecessary cloning. Use Vec::with_capacity for better perfomance
because we know the size of a file

* Use Vec::with_capacity where possible

for better perfomance

* Use `map_or` where possible

Instead of `map` and `unwrap` or an if statement

* Bump versions
Both plist_plus and rusty_libimobiledevice versions have been bumped