Fix map_or() usages & handle `sbservices_get_icon_pngdata` returning a null pointer  (#20)
* Fix `map_or()` dropping an underlying value

Using `map_or()` on an `Option` to get a pointer for a CString or a Plist results
in dropping the underlying value, thus a pointer is invalid. To prevent this
we use `as_ref()` to get an option with a reference to a value which won't be
dropped.

* Update Cargo.lock

* Handle `sbservices_get_icon_pngdata` returning a null pointer

By running some tests, I found out that the function may not return an error
and also leave an array uninitialized. Weather it's a bug or not, we need to
handle this in Rust, or else it causes a panic.